requirements.txt
fastapi
uvicorn
psycopg2-binary
xgboost
pandas
scikit-learn
numpy